chore: clean up unused packages and relocate test files
