+++ .gitignore
@@ -16,5 +16,7 @@ build
 .externalNativeBuild
 .cxx
 local.properties
-/app/build/
+app/build/
 build.log
+app/release
+/references/
